Moved dataverse logic to a separate class
